pdf_magic/requirements.txt
PyQt5
PyPDF2
pdf2image
Pillow
pytesseract
python-docx
pdf2docx
pikepdf
pdfminer.six